requirements.txt
streamlit>=1.32.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.1.0
apscheduler>=3.10.0
cryptography>=41.0.0
pytz
plotly>=5.17.0
numpy>=1.24.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
feedparser>=6.0.0
schedule>=1.2.0
watchdog>=4.0.0
streamlit-autorefresh
